[N/A] NavController 코드 정리, PagingModule 제거
