+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/ui/main/MainNavController.kt
@@ -3,24 +3,24 @@ package nlab.practice.jetpack.ui.main
 import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
-import nlab.practice.jetpack.util.nav.AbsNavController
+import nlab.practice.jetpack.util.nav.findFragmentByTag
 
 /**
  * Main 에서 사용하는 NavController 정의
  *
  * @author Doohyun
  */
-class MainNavController(fm: FragmentManager, @IdRes resId: Int) : AbsNavController(fm, resId) {
+class MainNavController(val fragmentManager: FragmentManager, @IdRes val containerResId: Int) {
 
     inline fun replacePrimaryFragment(tag: String, crossinline fragmentProvider: () -> Fragment) {
         val fragmentTransaction = fragmentManager.beginTransaction()
 
-        val targetFragment: Fragment = fragmentManager.findFragmentByTag(tag, fragmentTransaction, fragmentProvider)
+        val targetFragment: Fragment =
+                fragmentManager.findFragmentByTag(containerResId, tag, fragmentTransaction, fragmentProvider)
 
         // 현재 Primary Nav 가 보여야할 Target 과 다르면 숨김처리 수행
         if (fragmentManager.isNeedChangePrimaryNavigationFragment(targetFragment)) {
             fragmentManager.primaryNavigationFragment?.run { fragmentTransaction.hide(this) }
-
 
             /**
              * NOTE
@@ -34,5 +34,12 @@ class MainNavController(fm: FragmentManager, @IdRes resId: Int) : AbsNavControll
                     .setReorderingAllowed(true)
                     .commitNowAllowingStateLoss()
         }
+    }
+
+    /**
+     * primaryNavigationFragment 를 변경해야하는가?
+     */
+    fun FragmentManager.isNeedChangePrimaryNavigationFragment(targetFragment: Fragment): Boolean {
+        return primaryNavigationFragment?.let { it != targetFragment } ?: true
     }
 }
