+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/recyclerview/AbsGenericItemAdapter.kt
@@ -4,7 +4,7 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
 /**
- * Header, Footer,
+ * Header, Footer, Content 를 지원하는 어댑터
  *
  * @author Doohyun
  */
